==> 1-train-model.ipynb <==
import pickle

with open("texts.pkl",'rb') as f:
    texts  = pickle.loads(f.read())
with open("all_label_bio.pkl",'rb') as f:
    labels  = pickle.loads(f.read())
print(len(texts), len(labels))

import torch  
from torch.utils.data import DataLoader, TensorDataset  
from transformers import BertTokenizer, BertForTokenClassification, AdamW, get_linear_schedule_with_warmup  
from tqdm import tqdm

dic = {"O":0, "B-LOC":1, "I-LOC":2}  
  
# initialize the tokenizer
tokenizer = BertTokenizer.from_pretrained('../bert-base-uncased')  
  
# turn texts into input_ids, attention_masks, label_ids
input_ids = []  
attention_masks = []  
label_ids = []  
  
max_seq_length = 100 
  
for text, label in tqdm(zip(texts, labels)):  
    encoded_text = tokenizer(text, return_tensors='pt', padding='max_length', truncation=True, max_length=max_seq_length)  
    input_ids.append(encoded_text['input_ids'].squeeze(0))  
    attention_masks.append(encoded_text['attention_mask'].squeeze(0))  
      
    # turn label into ids 
    encoded_label = [0]+[dic.get(l, 0) for l in label] 
    encoded_label += [0] * (max_seq_length - len(encoded_label)) 
    encoded_label = encoded_label[:100] 
    encoded_label = torch.tensor(encoded_label, dtype=torch.long)  
    label_ids.append(encoded_label)  
from torch.utils.data import random_split
# turn lists into tensors
input_ids = torch.stack(input_ids, dim=0)  
attention_masks = torch.stack(attention_masks, dim=0)  
label_ids = torch.stack(label_ids, dim=0) 

# create dataset
dataset = TensorDataset(input_ids, attention_masks, label_ids)

# Print the first few samples of the dataset
for i in range(5):
    input_id = dataset[i][0]
    attention_mask = dataset[i][1]
    label_id = dataset[i][2]
    
    print(f"Sample {i + 1}:")
    print("Input IDs:", input_id)
    print("Attention Mask:", attention_mask)
    print("Label IDs:", label_id)
    print()

from torch.utils.data import DataLoader, random_split

# Split the dataset into training and testing sets
train_size = int(0.8 * len(dataset))
test_size = len(dataset) - train_size
train_dataset, test_dataset = random_split(dataset, [train_size, test_size])
print(len(train_dataset), len(test_dataset))

# Create data loaders
train_data_loader = DataLoader(train_dataset, batch_size=32, shuffle=True)
test_data_loader = DataLoader(test_dataset, batch_size=32, shuffle=False)


# BERT Trainning

import torch
from transformers import BertTokenizer, BertForTokenClassification, AdamW, get_linear_schedule_with_warmup
from torch.utils.data import DataLoader, TensorDataset, random_split
from sklearn.metrics import classification_report

# load the pre-trained model
model = BertForTokenClassification.from_pretrained('../bert-base-uncased', num_labels=3)
device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
model = model.to(device)

# define the optimizer
loss_fn = torch.nn.CrossEntropyLoss().to(device)
optimizer = AdamW(model.parameters(), lr=1e-4, eps=1e-8)

# define the scheduler
epochs = 3
total_steps = len(train_data_loader) * epochs
scheduler = get_linear_schedule_with_warmup(optimizer, num_warmup_steps=0, num_training_steps=total_steps)


from torch.nn import functional as F  

def test():
    # Evaluation loop
    model.eval()
    all_labels = []
    all_preds = []
    with torch.no_grad():
        for batch in tqdm(test_data_loader):
            b_input_ids = batch[0].to(device)
            b_input_mask = batch[1].to(device)
            b_labels = batch[2].to(device)

            outputs = model(b_input_ids, attention_mask=b_input_mask, labels=b_labels)
            batch_predictions = outputs.logits
            # catch the predicted ids
            predicted_ids = torch.argmax(F.log_softmax(batch_predictions, dim=2), dim=2) 

            loss = outputs.loss

            # Only keep predictions for non-padding tokens
            for i, mask in enumerate(b_input_mask):
                true_len = torch.sum(mask).item()
                all_labels.extend(b_labels[i, :true_len].cpu().numpy())
                all_preds.extend(predicted_ids[i, :true_len].cpu().numpy())
                # print(len(all_labels), len(all_preds))

    print(f'Test Loss: {loss.item():.4f}')
    # Compute evaluation metrics
    report = classification_report(all_labels, all_preds, digits=4)
    print(report)
    return loss.item()

best_loss = test()
print('test loss before train:', best_loss)

# Training loop
for epoch in range(epochs):
    model.train()
    # Training loop for one epoch
    for batch_idx, batch in enumerate(train_data_loader):
        b_input_ids = batch[0].to(device)
        b_input_mask = batch[1].to(device)
        b_labels = batch[2].to(device)

        optimizer.zero_grad()
        outputs = model(b_input_ids, attention_mask=b_input_mask, labels=b_labels)
        loss = outputs.loss
        loss.backward()

        optimizer.step()
        scheduler.step()

        if (batch_idx + 1) % 100 == 0:
            print(f'Epoch [{epoch+1}/{epochs}], Step [{batch_idx+1}/{len(train_data_loader)}], Loss: {loss.item():.4f}')

    print(f'Epoch [{epoch+1}/{epochs}], Loss: {loss.item():.4f}')

    now_test_loss = test()
    # Check if current loss is the best
    if now_test_loss < best_loss:
        best_loss = now_test_loss
        # save the best model
        print('save model')
        torch.save(model.state_dict(), 'bert_ner_model.pt')
        
    model.train()

print('Best loss:', best_loss)
