==> 2-predict.ipynb <==
# Predict

def find_b_loc_i_loc_sequences(lst):  
    sequences = []  
    current_sequence = None  
    for i, tag in enumerate(lst):  
        if tag == 'B-LOC' and current_sequence is None:  
            current_sequence = [i]  # Start a new sequence  
        elif tag == 'I-LOC' and current_sequence is not None:  
            current_sequence.append(i)  # Append to the current sequence  
        elif tag != 'I-LOC' and current_sequence is not None:  
            sequences.append((current_sequence[0], current_sequence[-1]))  # Save the sequence  
            current_sequence = None  # Reset the current sequence  
    # Check if there is a final sequence to addif current  
    if current_sequence is not None:  
        sequences.append((current_sequence[0], current_sequence[-1]))  
    return sequences  

import torch  
from transformers import BertTokenizer  
from torch.nn import functional as F  
from transformers import BertTokenizer, BertForTokenClassification

# Load the saved model  
tokenizer = BertTokenizer.from_pretrained('../bert-base-uncased')  
model = BertForTokenClassification.from_pretrained('../bert-base-uncased', num_labels=3)  
max_seq_length = 100

# Load the pretrained model parameters  
model.load_state_dict(torch.load('bert_ner_model.pt'))  

# Set the model to evaluation mode  
model.eval()

# New text for prediction
test_texts = [
    'Finsbury Park is a friendly melting pot community composed of Turkish, French, Spanish, Middle Eastern, Irish and English families. <br />We have a wonderful variety of international restaurants directly under us on Stroud Green Road. And there are many shops and large Tescos supermarket right next door. <br /><br />But you can also venture up to Crouch End and along Greens Lanes where there will endless choice of Turkish and Middle Eastern cuisines.'
]

# Convert text to token IDs
input_ids = []
attention_masks = []

for text in test_texts:
    encoded_text = tokenizer(text, return_tensors='pt', padding='max_length', truncation=True, max_length=max_seq_length)
    input_ids.append(encoded_text['input_ids'].squeeze(0))
    attention_masks.append(encoded_text['attention_mask'].squeeze(0))

input_ids = torch.stack(input_ids, dim=0)
attention_masks = torch.stack(attention_masks, dim=0)

# If GPU is available, move the model and data to GPU
device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
model = model.to(device)
input_ids = input_ids.to(device)
attention_masks = attention_masks.to(device)

# Make predictions
with torch.no_grad():
    outputs = model(input_ids, attention_mask=attention_masks)
    predictions = outputs.logits

# Get predicted label IDs
predicted_ids = torch.argmax(F.log_softmax(predictions, dim=2), dim=2)

# Decode predicted label IDs to original labels
dic = {"O": 0, "B-LOC": 1, "I-LOC": 2}
inverse_dic = {v: k for k, v in dic.items()}

decoded_labels = []
for predicted_id in predicted_ids.cpu().numpy():
    decoded_label = [inverse_dic[id] for id in predicted_id]
    decoded_labels.append(decoded_label)


# Align labels with the original text
decoded_texts = []
for text, decoded_label in zip(test_texts, decoded_labels):
    word_tokens = tokenizer.tokenize(text)
    sequences = find_b_loc_i_loc_sequences(decoded_label[1:])
    end = []
    for j in sequences:
        if j[1] <= len(word_tokens):
            s = ''
            for idx in range(j[0], j[1] + 1):
                if '##' in word_tokens[idx]:
                    s += word_tokens[idx].replace('##', '')
                else:
                    s += ' ' + word_tokens[idx]
            end.append(s.strip())
    decoded_texts.append(end)
decoded_texts

# Apply to dataset

from tqdm import tqdm
import re

def clean_text(text):
    # Remove HTML tags
    text = re.sub(r'<[^>]+>', '', text)
    # Remove URLs
    text = re.sub(r'http\S+', '', text)
    # Remove special characters except for alphanumeric, spaces, periods, and commas. Note that we included \. and , in the exclusion set, so they will not be removed
    text = re.sub(r'[^\w\s.,]', '', text)
    # Remove numbers (if needed)
    # text = re.sub(r'\d+', '', text)
    # Replace multiple spaces with a single space
    text = re.sub(r'\s+', ' ', text)
    # Remove leading and trailing spaces
    text = text.strip()
    return text

def get_token(test_texts):
    # Convert text to token IDs
    input_ids = []
    attention_masks = []

    print('tokenizer')
    ok_test_texts = []
    for text in tqdm(test_texts):
        try:
            text = clean_text(text)
            encoded_text = tokenizer(text, return_tensors='pt', padding='max_length', truncation=True, max_length=max_seq_length)
            input_ids.append(encoded_text['input_ids'].squeeze(0))
            attention_masks.append(encoded_text['attention_mask'].squeeze(0))
            ok_test_texts.append(text)
        except:
            encoded_text = tokenizer('', return_tensors='pt', padding='max_length', truncation=True, max_length=max_seq_length)
            input_ids.append(encoded_text['input_ids'].squeeze(0))
            attention_masks.append(encoded_text['attention_mask'].squeeze(0))
            ok_test_texts.append('')
    
    input_ids = torch.stack(input_ids, dim=0)
    attention_masks = torch.stack(attention_masks, dim=0)
    return ok_test_texts, input_ids, attention_masks

def pre(ok_test_texts, input_ids, attention_masks, model):
    # If GPU is available, move the model and data to GPU
    device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
    model = model.to(device)
    input_ids = input_ids.to(device)
    attention_masks = attention_masks.to(device)

    # input_ids and attention_masks are your input data lists, each element is a batch of inputs
    batch_size = 32  # Define the batch size
    num_batches = len(input_ids) // batch_size  # Calculate the total number of complete batches

    # If the length of input data is not divisible by the batch size, consider the last batch
    if len(input_ids) % batch_size != 0:
        num_batches += 1

    # Perform batch predictions
    print('Performing predictions')
    all_decoded_texts = []
    with torch.no_grad():
        for i in tqdm(range(num_batches)):
            start_idx = i * batch_size
            end_idx = min((i + 1) * batch_size, len(input_ids))  # Ensure not to exceed the data range

            # Get the current batch of input data
            batch_test_texts = ok_test_texts[start_idx:end_idx]
            batch_input_ids = input_ids[start_idx:end_idx]
            batch_attention_masks = attention_masks[start_idx:end_idx]

            # Perform predictions
            outputs = model(batch_input_ids, attention_mask=batch_attention_masks)
            batch_predictions = outputs.logits

            # Get predicted label IDs
            predicted_ids = torch.argmax(F.log_softmax(batch_predictions, dim=2), dim=2)
            
            # Decode predicted label IDs to original labels
            dic = {"O": 0, "B-LOC": 1, "I-LOC": 2}
            inverse_dic = {v: k for k, v in dic.items()}
            
            decoded_labels = []
            for predicted_id in predicted_ids.cpu().numpy():
                decoded_label = [inverse_dic[id] for id in predicted_id]
                decoded_labels.append(decoded_label)
            
            # Align labels with the original text
            decoded_texts = []
            for text, decoded_label in zip(batch_test_texts, decoded_labels):
                word_tokens = tokenizer.tokenize(text)
                sequences = find_b_loc_i_loc_sequences(decoded_label[1:])
                end = []
                for j in sequences:
                    if j[1] < len(word_tokens):
                        s = ''
                        for idx in range(j[0], j[1] + 1):
                            if '##' in word_tokens[idx]:
                                s += word_tokens[idx].replace('##', '')
                            else:
                                s += ' ' + word_tokens[idx]
                        end.append(s.strip())
                decoded_texts.append(end)
            all_decoded_texts += decoded_texts
    return all_decoded_texts


test_texts = [
    'Finsbury Park is a friendly melting pot community composed of Turkish, French, Spanish, Middle Eastern, Irish and English families. <br />We have a wonderful variety of international restaurants directly under us on Stroud Green Road. And there are many shops and large Tescos supermarket right next door. <br /><br />But you can also venture up to Crouch End and along Greens Lanes where there will endless choice of Turkish and Middle Eastern cuisines.'
             ] 
ok_test_texts, input_ids, attention_masks = get_token(test_texts)
decoded_texts = pre(ok_test_texts, input_ids, attention_masks, model)
decoded_texts

import pandas as pd
csv_data = pd.read_csv('listings_d.csv')
text_Data = csv_data['neighborhood_overview'].tolist()

ok_test_texts, input_ids, attention_masks = get_token(text_Data)

decoded_texts = pre(ok_test_texts, input_ids, attention_masks, model)

csv_data['output'] = decoded_texts
csv_data.to_csv('listings_d_out.csv')

pd.set_option('display.max_colwidth', 3000) 

see_data = pd.read_csv('listings_d_out.csv')
see_data.dropna(subset=['neighborhood_overview', 'output'], inplace=True)
see_data[['neighborhood_overview','output']].head(30)
